# docs/ods_api_contract.yaml
openapi: 3.0.3
info:
  title: ODS Mogi das Cruzes — API
  description: >
    API do sistema de votação da 1ª Conferência Nacional ODS — Etapa Livre,
    realizada na Fatec Mogi das Cruzes nos dias 28/03 e 04/04/2026.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
  - url: https://api.ods-mogi.railway.app
    description: Produção (Railway)
...
paths:
...
  /propostas:
    post:
      tags: [Admin]
      summary: Cadastra uma nova proposta em um eixo existente
      description: >
        Rota protegida por um secret via header. Permite que a coordenação
        insira propostas de última hora diretamente no banco.
      parameters:
        - in: header
          name: X-Admin-Secret
          required: true
          schema:
            type: string
          description: Senha de administração configurada nas variáveis de ambiente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropostaInput'
            example:
              eixoId: 1
              titulo: "Nova Proposta de Urgência"
              descricao: "Descrição detalhada inserida no dia do evento..."
      responses:
        '201':
          description: Proposta criada com sucesso
        '401':
          description: Secret ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
              example:
                erro: nao_autorizado
                mensagem: Não autorizado. Credencial de admin inválida.
        '422':
          $ref: '#/components/responses/ValidacaoError'
  /propostas/{id}:
    delete:
      tags: [Admin]
      summary: Remove uma proposta (apenas se não houver votos)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: header
          name: X-Admin-Secret
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Removido com sucesso
        '400':
          description: Proposta já possui votos e não pode ser removida

# ─────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────
components:
  securitySchemes:
    sessaoToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token retornado no check-in ou cadastro. Enviado no header Authorization.
  responses:
...
    ValidacaoError:
      description: Dados inválidos ou campos obrigatórios ausentes
      content:
        application/json:
          schema:
            type: object
            properties:
              erro:
                type: string
                example: validacao
              campos:
                type: array
                items:
                  type: object
                  properties:
                    campo:
                      type: string
                    mensagem:
                      type: string
                example:
                  - campo: email
                    mensagem: E-mail inválido.
                  - campo: nome
                    mensagem: Nome completo é obrigatório.
  schemas:
    Erro:
      type: object
      properties:
        erro:
          type: string
        mensagem:
          type: string
...
    PropostaInput:
        type: object
        required:
          - eixoId
          - titulo
          - descricao
        properties:
          eixoId:
            type: integer
            example: 1
          titulo:
            type: string
            example: "Nova Proposta de Urgência"
          descricao:
            type: string
            example: "Descrição detalhada da proposta inserida pelo admin."